requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
ollama
numpy
pypdf
python-multipart
psycopg2-binary
sqlalchemy
pandas
python-dotenv
